+++ docs/changelog.rst
@@ -13,6 +13,7 @@ Changelog
 * Tools: added the ability to create temporary projects from the Projects panel.
 * Tools: added the ability to move objects to the camera view with ``Ctrl+Alt+F``.
 * Tools: imported models will now assign a default animation to imported state machines.
+* Tools: searching large projects in the Project Browser is now much quicker.
 * Runtime: added support for MP3 sound files.
 * Runtime: added triplanar mapping support to the mesh shader (+TRIPLANAR and +TRIPLANAR_LOCAL).
 * Runtime: animated meshes have now correct pre-computed bounds.